changed variables type in BetterProject1

+++ Projekt1/betterMatrixLib/include/matrixLib.h
@@ -21,14 +21,15 @@ void checkInput();
  * @param W - ilosc wierszy macierzy
  * @param K - ilosc kolumn w macierzy
  */
-void printMatrix(double **wynik, int W, int K);
+void printMatrix(float **wynik, int W, int K);
+
 /**
  * printMatrix - wyświtela wskazaną macierz
  * @param wynik -tablica, w ktorej zawarta jest macierz
  * @param W - ilosc wierszy macierzy
  * @param K - ilosc kolumn w macierzy
  */
-void printMatrix(int **wynik, int W, int K);
+void printMatrix(long int **wynik, int W, int K);
 
 /**
  * matrixAdd - dodaje 2 macierze
@@ -38,7 +39,8 @@ void printMatrix(int **wynik, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wynikowej
  */
-int **matrixAdd(int **a, int **b, int W, int K);
+long int **matrixAdd(long int **a, long int **b, int W, int K);
+
 /**
  * matrixAdd - dodaje 2 macierze
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -47,7 +49,7 @@ int **matrixAdd(int **a, int **b, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wynikowej
  */
-double **matrixAdd(double **a, double **b, int W, int K);
+float **matrixAdd(float **a, float **b, int W, int K);
 
 /**
  * matrixSubtract - odejmuje macierze
@@ -57,7 +59,8 @@ double **matrixAdd(double **a, double **b, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-int **matrixSubtract(int **a, int **b, int W, int K);
+long int **matrixSubtract(long int **a, long int **b, int W, int K);
+
 /**
  * matrixSubtract - odejmuje macierze
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -66,7 +69,7 @@ int **matrixSubtract(int **a, int **b, int W, int K);
  * @param K - ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-double **matrixSubtract(double **a, double **b, int W, int K);
+float **matrixSubtract(float **a, float **b, int W, int K);
 
 /**
  * matrixMultiply - mnozy macierze
@@ -77,7 +80,8 @@ double **matrixSubtract(double **a, double **b, int W, int K);
  * @param Kb - ilosc kolumn macierzy b
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-int **matrixMultiply (int **a, int **b, int Wa, int Ka, int Kb);
+long int **matrixMultiply(long int **a, long int **b, int Wa, int Ka, int Kb);
+
 /**
  * matrixMultiply - mnozy macierze
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -87,7 +91,7 @@ int **matrixMultiply (int **a, int **b, int Wa, int Ka, int Kb);
  * @param Kb - ilosc kolumn macierzy b
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-double **matrixMultiply (double **a, double **b, int Wa, int Ka, int Kb);
+float **matrixMultiply(float **a, float **b, int Wa, int Ka, int Kb);
 
 /**
  * matrixMultiplyByScalar -mnorzy macierz przez skalar
@@ -97,7 +101,8 @@ double **matrixMultiply (double **a, double **b, int Wa, int Ka, int Kb);
  * @param s - skalar
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-int **matrixMultiplyByScalar (int **a, int Wa, int Ka, double s);
+long int **matrixMultiplyByScalar(long int **a, int Wa, int Ka, long int s);
+
 /**
  * matrixMultiplyByScalar -mnorzy macierz przez skalar
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -106,7 +111,7 @@ int **matrixMultiplyByScalar (int **a, int Wa, int Ka, double s);
  * @param s - skalar
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-double **matrixMultiplyByScalar (double **a, int Wa, int Ka, double s);
+float **matrixMultiplyByScalar(float **a, int Wa, int Ka, float s);
 
 /**
  * matrixTranspoze - transponuje macierz
@@ -115,7 +120,8 @@ double **matrixMultiplyByScalar (double **a, int Wa, int Ka, double s);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-int **matrixTranspoze (int **a, int Wa, int Ka);
+long int **matrixTranspoze(long int **a, int Wa, int Ka);
+
 /**
  * matrixTranspoze - transponuje macierz
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -123,7 +129,7 @@ int **matrixTranspoze (int **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-double **matrixTranspoze  (double **a, int Wa, int Ka);
+float **matrixTranspoze(float **a, int Wa, int Ka);
 
 /**
  * matrixPower - podnosi macierz do wskazanej potegi
@@ -133,7 +139,8 @@ double **matrixTranspoze  (double **a, int Wa, int Ka);
  * @param st - stopien potegi
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-int **matrixPower (int **a, int Wa, int Ka, unsigned st);
+long int **matrixPower(long int **a, int Wa, int Ka, unsigned short st);
+
 /**
  * matrixPower - podnosi macierz do wskazanej potegi
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -142,14 +149,15 @@ int **matrixPower (int **a, int Wa, int Ka, unsigned st);
  * @param st - stopien potegi
  * @return - zwraca wskaźnik do pierwszej liczby w macierzy wyniowej
  */
-double **matrixPower (double **a, int Wa, int Ka, unsigned st);
+float **matrixPower(float **a, int Wa, int Ka, unsigned short st);
 
 /**
  * LUdecomposition -rozklad LU macierzy (do wyznacznika)
  * @param a -macierz do rozkladu
  * @param n -rozmiar macierzy (ilosc wierszy/kolumn - macierz musi byc kwadratowa)
  */
-bool LUdecomposition(int n, double **a);
+bool LUdecomposition(int n, float **a);
+
 /**
  * matrixDeterminant - liczy wyznacznik macierzy
  * @param a - tablica, w ktorej zawarta jest macierz a (jest to tablica typu double ze wzgledu na dzielenie w rozkladzie LU)
@@ -157,7 +165,7 @@ bool LUdecomposition(int n, double **a);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wyznacznik macierzy
  */
-double matrixDeterminant (double **a, int Wa, int Ka);
+float matrixDeterminant(float **a, int Wa, int Ka);
 //tylko double ze wzgledu na dekompozycje lu gdzie jest dzielenie
 
 /**
@@ -167,7 +175,8 @@ double matrixDeterminant (double **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return -zwraca watrosc "true" jesli macierz jest diagonalna lub wartosc "false" gdy nie jest diagonalna
  */
-bool matrixIsDiagonal(int **a, int Wa, int Ka);
+bool matrixIsDiagonal(long int **a, int Wa, int Ka);
+
 /**
  * matrixIsDiagonal - sprawdza czy macierz jest diagonalna
  * @param a -sprawdzana macierz
@@ -175,7 +184,7 @@ bool matrixIsDiagonal(int **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return -zwraca watrosc "true" jesli macierz jest diagonalna lub wartosc "false" gdy nie jest diagonalna
  */
-bool matrixIsDiagonal(double **a, int Wa, int Ka);
+bool matrixIsDiagonal(float **a, int Wa, int Ka);
 
 
 /**
@@ -183,13 +192,14 @@ bool matrixIsDiagonal(double **a, int Wa, int Ka);
  * @param a
  * @param b
  */
-void swap(int *a, int *b);
+void swap(long int *a, long int *b);
+
 /**
  * swap - zamienia dwie wartosci zmiennych miejscami
  * @param a
  * @param b
  */
-void swap(double *a, double *b);
+void swap(float *a, float *b);
 
 /**
  * sortRow - sortuje pierwszy rzad w macierzy (rosnaco)
@@ -197,14 +207,15 @@ void swap(double *a, double *b);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do podanej marcierzy z posortowanym pierwszym rzedem macierzy
  */
-int **sortRow(int **a, int Ka);
+long int **sortRow(long int **a, int Ka);
+
 /**
  * sortRow - sortuje pierwszy rzad w macierzy (rosnaco)
  * @param a - tablica, w ktorej zawarta jest macierz a
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do podanej marcierzy z posortowanym pierwszym rzedem macierzy
  */
-double **sortRow(double **a, int Ka);
+float **sortRow(float **a, int Ka);
 
 /**
  * matrixSortRows -sortuje wszystkie rzedy macierzy (rosnaco)
@@ -213,7 +224,8 @@ double **sortRow(double **a, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do marcierzy z posortowanymi rzedami
  */
-int **matrixSortRows (int **a, int Wa, int Ka);
+long int **matrixSortRows(long int **a, int Wa, int Ka);
+
 /**
  * matrixSortRows -sortuje wszystkie rzedy macierzy (rosnaco)
  * @param a - tablica, w ktorej zawarta jest macierz a
@@ -221,7 +233,7 @@ int **matrixSortRows (int **a, int Wa, int Ka);
  * @param Ka -ilosc kolumn w macierzy
  * @return - zwraca wskaznik do marcierzy z posortowanymi rzedami
  */
-double **matrixSortRows (double **a, int Wa, int Ka);
+float **matrixSortRows(float **a, int Wa, int Ka);
 
 //do zrobienia
 /**
@@ -232,7 +244,8 @@ double **matrixSortRows (double **a, int Wa, int Ka);
  * @param max - maksymalna wartosc do ktorej losowane beda liczby
  * @return - zwraca wskaznik do losowo wygenerowanej macierzy
  */
-int **matrixGenerate(int W, int K,int min,int max);
+long int **matrixGenerate(int W, int K, long int min, long int max);
+
 /**
  * matrixGenerate - generuje losowa macierz
  * @param Wa - ilosc wierszy w macierzy
@@ -241,5 +254,6 @@ int **matrixGenerate(int W, int K,int min,int max);
  * @param max - maksymalna wartosc do ktorej losowane beda liczby
  * @return - zwraca wskaznik do losowo wygenerowanej macierzy
  */
-double **matrixGenerate (int W, int K, double min, double max);
+float **matrixGenerate(int W, int K, float min, float max);
+
 #endif //PROJEKT1_MATRIXLIB_H
